requirements.txt
CairoSVG==2.0.1
Django==1.10.5
Pillow==4.0.0
Pyphen==0.9.4
WeasyPrint==0.36
appdirs==1.4.2
cairocffi==0.8.0
cffi==1.9.1
cssselect==1.0.1
django-betterforms==1.1.4
django-formtools==2.0
django-localflavor==1.4.1
html5lib==0.999999999
lxml==3.7.3
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
pycparser==2.17
pyparsing==2.1.10
pytz==2016.10
reportlab==3.4.0
six==1.10.0
tinycss==0.4
webencodings==0.5